Draft PR to track pending changes. I'm making individual PRs for each commit in this branch.

Changes:

* Added documentation. Including documentation of method runtime.
* Optimize internals such that unsetting a bit doesn't turn all operations into `O(bits)` time.
* Make the bitfield receiver a pointer (there's no point in copying this stuff around).
* Optimize IsEmpty to make it `O(1)`.
* Benchmark.