@Stebalien 2020-06-29
There's a bit of added complexity for tiny (size 1) bitfields, but that shouldn't matter in practice.